requirements.txt
flask
python-dotenv
line-bot-sdk>=3.0.0
openai
pinecone
